Add support for JavaScript, CSS, and JSON content models
The addition of the Scribuntu (Lua) content model is saved for a
separate patch. Conflicts with Extension:CodeEditor are controlled with
$wgCodeMirrorContentModels and $wgCodeEditorContentModels
(see I5de9b6a5ea). These are both temporary feature flags.

Just as with wikitext, other content models do not require WikiEditor to
use CodeMirror. For WikiEditor users however, we do the same as
Extension:CodeEditor and hide inapplicable toolbar buttons with CSS.
Buttons are then added for indentation, whitespace, line wrapping, and
searching to make the experience similar to CodeEditor. A new button is
added for toggling on and off autocompletion.

Additional default extensions for non-wikitext include:
* closeBrackets
* highlightSelectionMatches with wholeWords option, and custom styling

Dark mode support to be added in a subsequent patch.

Other changes:
* Add CodeMirrorPreferences#toggleExtension() for internal use
* Update i18n messages for preferences accordingly.

Bug: T373711
Change-Id: I0ea9e0d68521068dbec7b520c70d8b80211e02fc