CodeMirror: finalize stable interface for consumption by integrations
Move toggling logic to parent class, add life-cycle methods, and
harmonize the CodeMirror and CodeMirrorWikiEditor interfaces.

For toggling, it is better to keep the EditorView instance alive rather
than to rebuild completely. Essentially, we treat the `state` property
as the indicator of whether CodeMirror is 'enabled' instead of `view`.
When toggling off and then back on, we only need to create a new
EditorState. This will also re-init all extensions and redraw the
editor, whilst avoiding side effects like the re-addition of mw.hook
handlers.

CodeMirrorWikiEditor was written to be based on the old
ext.CodeMirror.WikiEditor.js file. Rework the class to utilize the new
initialize/activate/deactivate/destroy life cycle of the parent
CodeMirror class. The old methods are removed as they are not yet used
in any production code or on-wiki code (and the CM6 migration project
is technically not yet complete, anyway).

Move syncing of the scroll position, selection and focus area to the
CodeMirror class as this is not WikiEditor-specific.

With only one remaining hook that emits a jQuery object, and that one is
marked as deprecated, we rework the interface to mainly use a
HTMLTextAreaElement, though a $textarea property is kept for easier
integration with the jQuery-based WikiEditor and textSelection.

The Jest config is changed to share coverage reports. Thresholds for
failing are set based on roughly the current coverage level, but may be
changed as needed. For now, the intention is merely to encourage better
test coverage moving forward.

Stable interface changes:
* CodeMirrorWikiEditor#addCodeMirrorToWikiEditor has been removed in
  favor of initialize().
* CodeMirrorWikiEditor#enableCodeMirror has been removed in favor of
  activate().
* CodeMirrorWikiEditor#updateToolbarButton has been removed. The OOUI
  button already has its initial value set and a change listener to
  sync the active state of the button.
* CodeMirrorWikiEditor#switchCodeMirror has been removed in favor of
  toggle().
* The 'ext.CodeMirror.initialize' hook may return the VE surface as a
  second argument to the handler.
* The 'ext.CodeMirror.ready' and 'ext.CodeMirror.destroy' hooks
  return a HTMLTextAreaElement instead of a jQuery object.
* The 'ext.CodeMirror.ready' hook only returns the CodeMirror instance.
* The 'ext.CodeMirror.switch' hook *is* being used by production and
  on-wiki code, and so has been marked as deprecated.
* CodeMirror#logUsage is no longer marked as stable; it will soon be
  rewritten anyway (T373710).

Stable interface additions:
* CodeMirror#toggle, stable to use and override
* 'ext.CodeMirror.toggle' hook, stable to use
* CodeMirror#activate, stable to call and override by subclasses
* CodeMirror#deactivate, stable to call and override by subclasses
* CodeMirror constructor, stable to call and extend
* CodeMirror#defaultExtensions, stable to call and override
* CodeMirror#applyExtension is added as an easier means to dispatch a
  new state effect for an extension. This isn't yet used by our code.
  Stable to call.

Bug: T373720
Change-Id: I7173b1120a632a3a57034f419be4a088f1d86ad0