CodeMirrorModeMediaWiki: fix very long lines (continued)
In practice, when there is a very long line where the CM parsing stops early with an extension tag (e.g., `<ref>`) still opening, usually the extension tag is closed on the same line. This patch makes this assumption and pops out the `eatExtTagArea` tokenizer.

This patch also addresses T360566#10419378.

Bug: T360566
Bug: T366035
Change-Id: I01a78e9b10d1b3426f1accdd64e52d0de93369b7