Add metrics instrumentation using the VisualEditorFeatureUse schema
The VisualEditorFeatureUse (VEFU) schema is considered legacy to Event
Platform <https://wikitech.wikimedia.org/wiki/Event_Platform>, but is
still being used and is the recommended schema to leverage for metrics
strictly related to editing. A new feature called 'codemirror' is used
to lump CM-specific feature use. Possible actions include:

* activated – when CodeMirror is activated by any means
* deactivated – when CodeMirror is toggled off
* search – when the CodeMirrorSearch panel is opened
* keymap – when the keyboard shortcut dialog is opened
* prefs-display – when the CodeMirrorPreferences panel is opened
* prefs-activeLine – when the activeLine extension is applied
* prefs-bidiIsolation
* prefs-bracketMatching
* prefs-lineNumbering
* prefs-lineWrapping
* prefs-specialChars
* prefs-codeFolding
* prefs-autocomplete
* prefs-openLinks

New public and internal JS hooks are added to facilitate this:

* ext.CodeMirror.preferences.apply – fired when a CodeMirror preference
  is enabled or initially applied.
* ext.CodeMirror.keymap (internal) – the keymap help dialog is opened.
* ext.CodeMirror.search (internal) – the search panel is opened.

For now, the base CodeMirror class expects subclasses to implement
logEditFeature(). In a future patch, we may add our own logging for
standalone CodeMirorr usage (which could replace or supplement the VEFU
schema).

CodeMirrorVisualEditor#setupFeatureLogging() exists accordingly to trim
the fat, since the CodeMirror keymap and search features aren't
available in VE.

Logging to the old 'codemirrorusage' schema is removed for both CM5 and
CM6, as the dashboard has been retired for some time and the metrics
were never migrated to statslib (T359461).

Bug: T373710
Change-Id: I5fe28056e1665a5bd16fbce723ec82e168b812eb