CodeMirrorVisualEditor: new subclass for the 2017 editor integration
Move VE 2017 editor logic to a CodeMirror subclass using the same
life-cycle methods. This means the 2017 editor now shares the same
CodeMirror instance when toggling CM on and off.

This reworking doesn't completely remove knowledge of VE from the parent
CodeMirror class, but takes steps in that direction in an effort to
maintain separation of concerns.

The ES5-based OOUI classes remain as separate files, with the entrypoint
ve.ui.CodeMirrorAction being responsible for CodeMirror initialization.

Add Selenium tests guarding against various issues we've had with
CodeMirror + 2017 editor in the past. Also remove the word 'legacy' from
non-2017 editor tests, as they are not in fact legacy.

Bug: T357482
Change-Id: I101c33b90d2a1e547b21aeccc8ab294cd6ab1e4a