CodeMirror: use the EditorView's state property on form submission
CodeMirror#state is reliable to determine if CM is currently "activated"
since we (un)set it ourselves. However once it's applied with
EditorView#setState(), it appears it may be mutated by CM internals.
The code should be reworked to use CodeMirror#state only for state
memory and not persistence. For now, go to using the state property on
the EditorView as was done prior to I7173b1120a.

Add test asserting the textarea is properly updated.

Bug: T387253
Change-Id: Ieb52d589bc827c9f28db7f1edea83d466e3540ea