CodeMirror: don't persist EditorState after deactivation
When transactions (such as typing) are dispatched, a new EditorState
instance is created, but the original object is not updated. This is
what led to T387253. The CodeMirror#state property was only meant to be
a shorthand to CodeMirror#view.state anyway, so do away with it and
make the isActive getter into an actual property that's updated by the
the lifecycle methods.

Bug: T387253
Change-Id: I25b82319029d835e906e009aa8fc62624b9a597b