Merge namespace configs to DefaultPreferences, add support for model
Simplify namespace-related config settings. The intention is for the end
user to always have the ability to override sysadmin-imposed defaults.

$wgCodeMirrorDefaultPreferences accepts a boolean (on or off), or an
array containing any of:
* namespace IDs (integer) representing which namespaces the feature
  should be enabled by default.
* content models (string) for which content models the feature should be
  enabled in by default.

Add CodeMirrorPreferences#getDefaultPreferences() to compute and cache
a copy of $wgDeCodeMirrorfaultPreferences respecting the aforementioned
rules.

For now, $wgCodeMirrorLineNumberingNamespaces is left intact for CM5
compatibility. The config setting is however removed from ResourceLoader
which shaves off a tiny bit from the size of the startup module.

This patch also fixes an issue where if the preferences match the
default, we didn't delete the user option from the db as we should have.

Bug: T383550
Change-Id: I23998bd9da97b04de79b340c22fffff5b7622bf2