CodeMirror: bubble keyup, keydown and scroll events to the textarea
The goal here is to allow for more seamless integrations moving forward
(usually without the need for a hook at all), while not breaking
integrations that rely on the ext.CodeMirror.switch hook, or its
intended replacement ext.CodeMirror.toggle.

As part of this, we call toggle() from initialize(). This allows clients
to rely only on the ext.CodeMirror.toggle hook, making for an easier
migration that works in both CM5 and CM6. The logic on when to fire the
toggle hook is adjusted accordingly.

Clean up the init script which unnecessarily called
CodeMirror#setCodeMirrorPreference() directly.

Bug: T373720
Change-Id: I3a1716c5e795d3d3ee4cab41995767c59a1088eb