CodeMirrorKeymap: Make Ctrl-y work for Undo on all platforms
T365072 was intended to add Ctrl-Shift-z as an *additional* key binding
for Windows, but still have Ctrl-Y work. This fixes that by adding
support for multiple CodeMirrorKeyBindings for a single entry in the
keymapHelpRegistry, and using the original KeyBindings from upstream but
adding Ctrl-Shift-z for Windows.

Multiple CodeMirrorKeyBindings was already supported when generating the
Extensions in the #extension getter, but it was not supported in the
help dialog. Now any additional applicable key bindings are shown as
aliases in the same <kbd> element.

Bug: T365072
Bug: T380913
Change-Id: Ief4e996d9e38a2c8b95ba32bf788471f73f5cbad