CodeMirrorMediaWikiKeymap: give autocomplete a section in help dialog
When toggling preferences, the corresponding section in the help dialog
is supposed to be shown/hidden accordingly. This wasn't happening for
autocompletion because it didn't have its own section, so only the
key binding called 'autocomplete' was hidden, while 'startcompletion'
incorrectly stayed displayed. Grouping them into a section fixes this.

CodeMirrorKeymap wants the section name to match the preference name. As
a result, we need to rename i18n keys as codemirror-keymap-autocomplete
was already being used. Key changes like this are to be avoided, but
this is easier than changing the preference name - which would require a
maintenance script to update user preferences already stored in the db.

For a future patch: Show whichever relevant autocompletion keys still
apply outside of MediaWiki.

Change-Id: Iaafb7da86e411e42e457d4912e1386a0075b1705