CodeMirrorExtensionRegistry: container class for Extension management
Decouple extension management from CodeMirrorPreferences into a
standalone container class that can be passed around as needed. This is
also intended as an abstraction layer over CM internal Extension
management. Changing of preferences is now reserved for internal use
only.

The constructor is private and for use only by the CodeMirror class. The
methods however are encouraged as a more developer-friendly alternative
to CodeMirror#applyExtension() if the Extension is to be reconfigurable.

Remove the 'ext.CodeMirror.preferences.change' hook as redundant to
'ext.CodeMirror.preferences.apply'.

In the Jest tests, we now emit console warnings, and suppress them only
when they are expected.

Bug: T389441
Change-Id: I800ff76d502ba43236a3c998a7813940085c14d0