CodeMirrorSearch: force commit() when open search panel
Native CM will clear the replace input every time the search panel is
opened (probably a bug of its own). This seems to cause a mismatch
between our SearchQuery object and the one used by the internal
SearchState, which is not exported. Work around this by calling commit()
when the search panel is opened, and test against getSearchQuery().

Other changes:
* With CodeMirrorSearch now accessible, mark more methods as private
* Add test case for toggle buttons

Bug: T392343
Change-Id: Id0b4057746059fed940f273d9ef3ddf79f70c735