Add support for multiple CM instances, add to Special:ExpandTemplates
Introduce CodeMirrorChild (could probably be better named) that takes a
'primary' CodeMirror instance and keeps the preferences and toggle state
synced. This upstreams what was done in Ie937f4e8cb.

Rework Hooks to take an array of additional textareas. For now, these
are assumed to be the same mode as the primary instance. At a later
time, we can add support for multiple modes. This would be useful for
Special:ExpandTemplates since it can also output XML/HTML (we'd need to
also add HTML as a supported mode, however).

Special:ExpandTemplates uses a required textarea. Since the textarea is
hidden while CM is active, the browser will complain with "invalid form
control … is not focusable" on submission. For now, we work around this
by simply removing the 'required' attribute and leaving the burden to
implement validations on the CodeMirror client. In the case of
Special:ExpandTemplates, there is backend validation. A TODO is left to
improve this later.

Additionally, sync focus state during initialization before the textarea
is first hidden.

Bug: T384148
Change-Id: Id687f1f3f0316210605d91c942786127e2aed2a2