Remove ContentModels ext attribute, make config flag associative
CodeMirrorContentModels was meant to be a simple way to assign a CM mode
to a non-standard content model. ProofreadPage is the only user of this
attribute. To make the system less confusing, drop the ContentModels
attribute entirely and force extensions to use the
CodeMirrorGetModeHook, similar to how CodeEditor integration works.

proofread-page is now set in $wgCodeMirrorContentModels. This is a
temporary config flag used to control conflicts with CodeEditor, so it's
OK to reference other extensions here as we'll eventually remove the
setting anyway.

Other changes:
* Make $wgCodeMirrorContentModels an associated array to make it easier
  to change default values to false.
* Reorder properties in Hooks to match constructor signature
* Clean up Hooks::getMode() - it only needs the Title, and switch is a
  nicer control structure than a bunch of elseifs.
* Move hard-coded list of supported modes to a constant, and correct
  error message
* Fix tests that should have failed as previously written

Bug: T393742
Follow-Up: I0ea9e0d68521068dbec7b520c70d8b80211e02fc
Change-Id: Ie737c205d5af6a7c656229048a2844b397d89ab4