CodeMirrorTextSelection: don't dispatch events on readonly textareas
With the EditorState#readOnly facet set, EditorView#dispatchEvent()
is supposed to do nothing. This suggests a possible bug in the upstream
library. EditorView#editable is also available but that would prevent
the editor from being focused – something we want to allow for readonly
textareas (for code folding, search, etc.).

For now, we will check against the readOnly property and not dispatch
changes when it is set. This doesn't prevent users from dispatching
changes with the CodeMirror#view property, but CodeMirrorTextSelection
is the canonical interface for making changes to the document, so the
readOnly property should be respected during normal user interaction.

Change-Id: Id1859e017ee43c64f8d7c0c98b560df5ce2f1475