config: enable linting of Scribunto namespace by default
We want to avoid hard-coding things that depend on a different
extension, but in this case we need to change site configuration.
Programmatic changes to configuration is discouraged,
see https://w.wiki/EjHP

It's better to just hard-code it here in CodeMirror than require every
installation to set:

$wgCodeMirrorDefaultPreferences['lint'][] = 'Scribunto';

Even then it is assuming the default prefs for linting is in the array
form, when it can also be a boolean.

Bug: T373711
Change-Id: I2c750260fe03122b8b278bdac77dc7cfb631931d