Hooks, Preferences: use the CodeMirror mode in place of content model
Some extensions like JsonConfig can create arbitrary content models.
This causes a great deal of complexity for configuration such as
$wgCodeMirrorContentModels, which is used to piecemeal add new content
models in order to prevent conflicts with CodeEditor.

This patch reworks the system to use the CodeMirror "mode" (akin to the
"lang" in CodeEditor) as the canonical identifier for specifying
languages in the configuration. CodeMirrorContentModels is now
CodeMirrorEnabledModes, and CodeMirrorDefaultPreferences now treats
string[] as referring to the CM mode and not the content model.

This change will require updates to consumers of CodeMirrorGetModeHook.
While we're at it, let's change the interface to be more strict by using
a scalar return type.

Require PHP 8.1 and make use of new features.

Bug: T399659
Depends-On: I6c7d51530152449040dd27a696d51bbb0bc7de34
Change-Id: I038d0ee026e8f64b909d3a9fdf6bf57bd01f67a2