CodeMirrorPreferences: add link to panel to open all prefs in a dialog
The preferences panel is great for quickly seeing the effects of a
preference, but we can't cram all of them in there without consuming too
much real estate within the editor. This is especially a problem on
mobile.

Introduce $wgCodeMirrorPrimaryPreferences to specify which preferences
should go in the panel, and add an "Advanced" link that opens a dialog
with the full preferences. The chosen "primary" preferences are subject
to debate, but I went with ones I thought are more likely to be toggled
on and off, and have more noticeable effects.

The dialog is accessible directly with Alt-Shift-,

Within the dialog, we break the preferences out into fieldsets with
legends to make them more comfortable to navigate. The list is short
now, but is expected to grow (T393664#11069380).

Bug: T393664
Change-Id: I0830578f6f34ade50bf4b98c06eae94c2a5814fc