Use a LanguageSupport object for all modes; rework use of Hook JS vars
See [[mw:Special:Diff/7775359]]. We shouldn't rely on Hooks-provided
variables in code we expect to be used outside of action=edit etc.
Only the init script should know about them, and beyond that we pass
everything around akin to dependency injection. Leftover are uses of
cmDebug in the workers, which is fine for now given it won't error out.
We can use extCodeMirrorConfig wherever we want, as it runs in a static
context.

Remove jquery.client from lazy-loading module list in
ve.ui.CodeMirrorAction since it's already a hard dependency.

Change-Id: I076dc08b0890c57bfae108c230f69d7545876c55