Combine non-wikitext RL modules, and introduce CodeMirrorMode interface
To reduce overall ResourceLoader module count, we do the same as
Extension:CodeEditor and combine all the "major" languages into a single
module. As it happens, each language package shared a lot of the same
dependencies, so this seems like a good optimization overall.
ext.CodeMirror.v6.modes, including the Vue integration (added in a
subsequent patch), is roughly the same size as the CodeEditor equivalent
at ~850 kB uncompressed.

A new interface-like class, CodeMirrorMode, is added to enforce a
consistent implementation for each mode. Combined with
codemirror.mode.exporter.js, it follows the upstream CM convention of
language packages providing a method with the same name as the mode,
which returns a LanguageSupport instance (or in our case, a
LanguageSupport-compatible class).

Modes can be unpacked and used like:

    const { javascript } = require( 'ext.CodeMirror.v6.modes' );
    new CodeMirror( myTextArea, javascript() );

CodeMirrorMediaWiki still lives as a separate RL module due to its
high-traffic use, and unique dependencies and implementation, but is
changed to extend CodeMirrorMode for consistency.

Bug: T373711
Change-Id: Id5359f31e8497f81f9ca8a8d794a3e81d4ef96a7