CodeMirrorMediaWiki: export method with same name as mode per convention
It is convention for CodeMirror language packages to export a method
with the same name as the mode, and for that to return a
LanguageSupport. This patch makes ext.codemirror.mode.mediawiki behave
in the same way, though the old default export is still supported.

In preparation of Id5359f31e84

Change-Id: Ib031787d5ea595dcc4e64ce8cf948920680652c6