wikilint: revise to a relatively lax yet explorative rule config
While we're still in Beta is the time to test the waters. This still
leaves some rules disabled, but largely goes off of the upstream
default.

WikiLint doesn't support an "info" level (as is typical for linters) so
we use a workaround.

Minor performance optimization: Only give WikiLint the messages it
needs.

TODO: add missing messages for the new rules we're exposing. There
appears to be an existing issue with i18n and WikiLint so we can tackle
this in a separate patch.

Bug: T381577
Change-Id: Id0ce92e843b83689d2300d3f3793251efd3b4b63