preferences: add a means to lock a pref to a value and prevent changes
This will be useful for integrations where some preference and its
extension are incompatible or cause problems, particularly say with a
CodeMirrorChild instance.

Add CodeMirrorPreferences#lockPreference() to lock the extension config
to a certain value and disable the checkbox in the panel and dialog.

Also, make sure the CodeMirrorChild does not trigger redundant API
requests when changing a preference value.

Change-Id: I6bc6286411e045905dbfc632f7266302ec2dad0c