CodeMirrorMediaWiki: add pref to give <ref> content a shaded background
This reintroduces a similar background color to was added in I1e1f309e58
and later removed. In this patch, the color is slightly adjusted and
with a 50% opacity to avoid the conflict with ::selection styles. A dark
mode variant is also custom tuned. Both are subject to change based on
feedback.

The preference is enabled by default.

Bug: T384908
Change-Id: I44bc0490378b87366038d2e805163adf545827bc