Add preference to auto-fold all <ref> tags on page load
On large articles like [[w:Barack Obama]] we are able to fold all ref
tags within 1 second, so the belief is this may be acceptably fast for
the vast majority of articles.

On page load, the slowness is barely noticeable, but it's a little more
evident when turning on the preference for the first time in the dialog.
Barring objections, I still think this is worthwhile given how useful
the feature is and how popular it may become. Users will likely be
content with a one-off, slight delay.

* Adds a new section "References" to the preferences dialog
* A new preference, 'foldAllRefs' (default off), will attempt to fold
  all <ref> tags in the document when it is first loaded.
* Add a 'slow' boolean argument to #registerCallback to visually
  indicate foldAllRefs is "potentially slow". This same option may be
  used as a disclaimer on other preferences, too, so it's also now
  available for #registerExtension, and likewise for the
  CodeMirrorKeyBinding type definition.
* The existing 'foldRef' keyboard shortcut now always uses foldAllRefs
  with ensureSyntaxTree(). As an opt-in, on-demand feature, it makes
  sense to parse the whole tree, I think. Non-wikitext code can be much
  bigger, so it makes sense to keep it using the known syntax tree.
* A timeout of 1 second is placed, and if exceeded, will emit a warning
  and fallback to using the known syntax tree.
* Use the presence of <ref> in the DataScript's 'tags' property as the
  indicator of whether Extension:Cite is installed.

Bug: T407824
Change-Id: Icdf8536707b212b41e2c72e8d8721b9b01de679d