CodeMirrorVisualEditor: add integration with DiscussionTools
DiscussionTools does not have knowledge that a 'codeMirror' button
should exist in its toolbar, so we need to dynamically add it during DT
initialization, and retroactively run the VE Command for CodeMirror to
take effect.

The line numbering extension in DiscussionTools is disabled as it's not
as useful for discussions and unnecessarily consumes UI real estate on
mobile.

Other targets can opt-in by adding 'codeMirror' in their toolbar config.
See I3054a1ca92 for an example of this.

Remove presumptive CSS rules about the padding and instead compute it at
run time, guaranteeing it matches the skin and for any VE target.

Move setting of the 'usecodemirror' preference to ve.ui.CodeMirrorTool
from CodeMirrorVisualEditor. This prevents the preference being
enabled/disabled when switching between source and visual mode. We do
however want to destroy the CM instance when changing to visual, since
that erases all of the DOM and may otherwise leave unwanted artifacts
behind such as hook handlers.

Desktop article targets can reliably be loaded before VE initializes
with the targetLoader's with #addPlugin method.

Bug: T407918
Change-Id: Ic404c15727234cac3ec2adf6a1513f9104550629