CodeMirrorWikiEditor: add style module to prevent FOUCs
This is the CodeMirror equivalent to T73015 (I0ce663ebad). When editing
code (non-wikitext), a larger edit window is desired to be able to scan
the code more easily. This patch sets the height to roughly 25 lines of
code, which is about the same as what CodeEditor does.

This approach doesn't work for read-only pages as there's no container
around the textarea for us to style. We still load the style module
however, as we may add more things to it later that we'd want to effect
all action=edit pages.

Change-Id: I1f866f6bc65494190dcac2e6a8db04712049c27c