performance: lazy load CM module when VE wikitext editing session begins
Remove ext.CodeMirror.v6 is a dependency of ext.CodeMirror.v6.visualEditor

Instead it should get loaded with ext.CodeMirror.v6.mode.mediawiki which
is lazy-loaded in ve.ui.CodeMirrorAction.v6.js

The loading order of the stylesheets means we need to use !important for
the background gutter or else CodeMirror styles will apply.

jQuery.client is used by CodeMirrorVisualEditor so we lazy-load that
dependency as well. It incidentally is also loaded by
ext.CodeMirror.v6.mode.mediawiki, but that may not always remain true.

Bug: T409066
Change-Id: Ice03e651207a7b72e0080389680a7271f3e193d1