CodeMirrorVisualEditor: make contenteditable inert
The inert attribute will disable user interaction on the element and its
descendants, which is precisely what we want in the 2017 editor. 'inert'
may not work in older browsers, and a fully accurate polyfill would be
non-trivial. A simple workaround is to use tabindex="-1" to prevent
tabbing to the CodeMirror contenteditable.

Bug: T412827
Change-Id: I1ce8cfe654ad08e6bc2e712511c32ba6e53233ff