Use toggle input boxes for buttons that control panels
The panels are toggleable, so using a toggle button is more fitting and
intuitive. A new internal mw.hook is added for gotoLine, but it and the
other hooks are only used as a means to trigger an update to the toggle
button. We check the editor state for the StateField to determine the
toggle state, ensuring we don't inadvertently react to another
CodeMirror instance (since mw.hook is global).

OOUI icons are copied to the local repo so that they use mask-image
instead of a background-image, allowing us to style them as progressive
in the active state. We also need to use names that differ from the
built-in icons, or the background-image will still be set.

This patch also removes the mask-size property from icons, thus
preventing them from being resized based on Vector-2022 accessibility
settings.

Change-Id: I899f8dc8486f093dfd897447c147a5d0e61603bf